backend/requirements.txt
asgiref==3.8.1
base58==2.1.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
ecdsa==0.19.0
pycryptodome==3.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
